requirements.txt
aiohttp ~= 3.12.15
discord.py[speed] ~= 2.6.4
python-dotenv ~= 1.1.1
tomlkit ~= 0.13.3
tzdata; os_name == "nt"
